x11: move x11 specific code to x11 from linux/mod.rs
